Revert "Atualização do README.md com melhorias na descrição do projeto e estrutura do repositório. Remoção de funcionalidades de seguir usuários e ajustes nas classes de perfil e comentários. Adição de novos testes para a funcionalidade de sair de comunidades e correções em menus. Ajustes na lógica de busca e contagem de curtidas."
This reverts commit 35f2520df7029fb8ae7bd10339684aa900534959.

+++ src/domain/comentario.cpp
@@ -13,9 +13,7 @@ void Comentario::curtir(int idUsuario) {
     else idsCurtidas.erase(it);
 }
 
-int Comentario::quantidadeDeCurtidas() const {
-    return static_cast<int>(idsCurtidas.size());
-}
+int Comentario::quantidadeDeCurtidas() const { return idsCurtidas.size(); }
 int Comentario::getId() const { return id; }
 int Comentario::getIdPost() const { return idPost; }
 int Comentario::idAutorObter() const { return idAutor; }
